third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-img-element/naturalWidth-naturalHeight-width-height.tentative-expected.txt

This is a testharness.js-based test.
Found 162 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] SVG image, no natural dimensions
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image with width/height attrs, no natural dimensions
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image with width/height attrs, no natural dimensions (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image with width attr, no natural dimensions
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image with width attr, no natural dimensions (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image with height attr, no natural dimensions
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image with height attr, no natural dimensions (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, percengage natural dimensions
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, percengage natural dimensions (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, negative percengage natural dimensions
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, negative percengage natural dimensions (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width of 0
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width of 0 (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height of 0
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height of 0 (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width being negative
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural width being negative (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural height being negative
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height being negative (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions, and aspect ratio from viewBox
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions, and aspect ratio from viewBox (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, percengage natural dimensions, and aspect ratio from viewBox
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, percengage natural dimensions, and aspect ratio from viewBox (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, negative percengage natural dimensions, and aspect ratio from viewBox
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width being negative, and aspect ratio from viewBox
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural width being negative, and aspect ratio from viewBox (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural height being negative, and aspect ratio from viewBox
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height being negative, and aspect ratio from viewBox (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width/height
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width/height (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 height
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 height (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width/height
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width/height (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 height
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 height (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width/height
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width/height (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 height
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 height (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width and height being negative, and aspect ratio from viewBox
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural width and height being negative, and aspect ratio from viewBox (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image with width/height attrs, no natural dimensions (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image with width/height attrs, no natural dimensions (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image with width attr, no natural dimensions (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image with width attr, no natural dimensions (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image with height attr, no natural dimensions (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image with height attr, no natural dimensions (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, percengage natural dimensions (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, percengage natural dimensions (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, negative percengage natural dimensions (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, negative percengage natural dimensions (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width (with srcset/1x)
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width (with srcset/1x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width of 0 (with srcset/1x)
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width of 0 (with srcset/1x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height of 0 (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height of 0 (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width being negative (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural width being negative (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural height being negative (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height being negative (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions, and aspect ratio from viewBox (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions, and aspect ratio from viewBox (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, percengage natural dimensions, and aspect ratio from viewBox (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, percengage natural dimensions, and aspect ratio from viewBox (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width being negative, and aspect ratio from viewBox (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural width being negative, and aspect ratio from viewBox (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural height being negative, and aspect ratio from viewBox (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height being negative, and aspect ratio from viewBox (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width/height (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width/height (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 height (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 height (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width/height (with srcset/1x)
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width/height (with srcset/1x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width (with srcset/1x)
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width (with srcset/1x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 height (with srcset/1x)
  assert_equals: naturalHeight expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 height (with srcset/1x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width/height (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width/height (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 height (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 height (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, with natural width and height being negative, and aspect ratio from viewBox (with srcset/1x)
  assert_equals: naturalWidth expected 300 but got 0
[FAIL] SVG image, with natural width and height being negative, and aspect ratio from viewBox (with srcset/1x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 300 but got 0
[FAIL] SVG image, no natural dimensions (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, no natural dimensions (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image with width/height attrs, no natural dimensions (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image with width/height attrs, no natural dimensions (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image with width attr, no natural dimensions (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image with width attr, no natural dimensions (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image with height attr, no natural dimensions (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image with height attr, no natural dimensions (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, percengage natural dimensions (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, percengage natural dimensions (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, negative percengage natural dimensions (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, negative percengage natural dimensions (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width (with srcset/2x)
  assert_equals: naturalHeight expected 75 but got 0
[FAIL] SVG image, with natural width (with srcset/2x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 75 but got 0
[FAIL] SVG image, with natural height (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural height (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width of 0 (with srcset/2x)
  assert_equals: naturalHeight expected 75 but got 0
[FAIL] SVG image, with natural width of 0 (with srcset/2x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 75 but got 0
[FAIL] SVG image, with natural height of 0 (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural height of 0 (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width being negative (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural width being negative (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height being negative (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural height being negative (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, no natural dimensions, and aspect ratio from viewBox (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, no natural dimensions, and aspect ratio from viewBox (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, percengage natural dimensions, and aspect ratio from viewBox (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, percengage natural dimensions, and aspect ratio from viewBox (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, negative percengage natural dimensions, and aspect ratio from viewBox (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width being negative, and aspect ratio from viewBox (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural width being negative, and aspect ratio from viewBox (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height being negative, and aspect ratio from viewBox (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural height being negative, and aspect ratio from viewBox (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width/height (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width/height (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 width (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 height (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, no natural dimensions, viewBox with 0 height (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width/height (with srcset/2x)
  assert_equals: naturalHeight expected 75 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width/height (with srcset/2x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 75 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width (with srcset/2x)
  assert_equals: naturalHeight expected 75 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 width (with srcset/2x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 75 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 height (with srcset/2x)
  assert_equals: naturalHeight expected 75 but got 0
[FAIL] SVG image, with natural width, viewBox with 0 height (with srcset/2x) (when not rendered)
  assert_equals: naturalHeight when not rendered expected 75 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width/height (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width/height (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 width (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 height (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural height, viewBox with 0 height (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
[FAIL] SVG image, with natural width and height being negative, and aspect ratio from viewBox (with srcset/2x)
  assert_equals: naturalWidth expected 150 but got 0
[FAIL] SVG image, with natural width and height being negative, and aspect ratio from viewBox (with srcset/2x) (when not rendered)
  assert_equals: naturalWidth when not rendered expected 150 but got 0
Harness: the test ran to completion.
